Reorg preprocessing module
