+++ erftools/preprocessing/__init__.py
@@ -1,2 +1,4 @@
-from .preprocessing import WRFInputDeck, LambertConformalGrid
+from .wrf_inputs import WRFInputDeck
+from .grids import LambertConformalGrid
 from .hrrr import NativeHRRR, hrrr_projection
+from .real import get_zlevels_auto
